Finished Corrections

+++ src/main/java/com/example/userscrud/controller/UserController.java
@@ -4,10 +4,12 @@ import java.net.URI;
 import java.util.List;
 
 import com.example.userscrud.exception.DuplicateUsersException;
+import com.example.userscrud.exception.UserNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
+import org.springframework.ui.ModelMap;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
@@ -39,13 +41,13 @@ public class UserController {
 		return userService.getAllUsers();
 	}
 	
-	@GetMapping("/{email}")
-	public User retrieveUser(@PathVariable String email) {
+	@GetMapping("/getByEmail")
+	public User retrieveUser(@RequestParam String email) {
 		return userService.getUser(email);
 	}
 	
-	@DeleteMapping("/{email}")
-	public void deleteUser(@PathVariable String email) {
+	@DeleteMapping("/deleteByEmail")
+	public void deleteUser(@RequestParam String email) {
 		userService.deleteUser(email);
 	}
 	
@@ -87,16 +89,25 @@ public class UserController {
 		return "This shit is working";
 	}
 
-	@DeleteMapping("/{name}")
-	public ResponseEntity<?> deleteByName(@PathVariable String name){
-		userService.deleteUserByName(name);
-		return ResponseEntity.noContent().build();
+	@RequestMapping("/deleteUserByName")
+	public ResponseEntity<List<User>> deleteByName(@RequestParam String name){
+		List<User> users = userService.getUserByName(name);
+		System.out.println("delete by name called");
+		if(users == null || users.isEmpty()){
+			throw new UserNotFoundException("\n There is no user with the name: " + name);
+		}
+		else if(users.size() > 1){
+			throw new DuplicateUsersException("Cannot delete this user because there is more than one user with this name.");
+		}
+		else {
+			userService.deleteUserByName(name);
+			List<User> remainingUsers = userService.getAllUsers();
+			return ResponseEntity.ok(remainingUsers);
+		}
 	}
 
 	@ExceptionHandler(DuplicateUsersException.class)
 	public ResponseEntity<String> duplicateUsersException(DuplicateUsersException e){
 		return ResponseEntity.status(HttpStatus.CONFLICT).body(e.getMessage());
 	}
-	
-
 }
